requirements.txt
gitpython
pyyaml
sigma-cli
pyyaml
